Lock event handler execution for nestedjoin
When adding collections that contain the same key to a nestedjoin, there
can be a race post-refactor where events for the same key can arrive out
of order. We solve this by simply locking the event handler execution
across the entire nested join. This probably reduces performance a bit
but for this approach, I believe it's necessary for correctness since
collection changes cause contention within the same handler context. The
state of the cache is always correct, but registered handler functions
can be executed out of order

Signed-off-by: Keith Mattix II <keithmattix@microsoft.com>